% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{asQualitative}
\alias{asQualitative}
\title{Convert a data frame containing integers to a qualitative data set with factors.}
\usage{
asQualitative(x)
}
\arguments{
\item{x}{a vector or a matrix or a data frame}
}
\value{
a qualitative data set with factors
}
\description{
Convert a data frame containing integers to a qualitative data set with factors.
}
