\docType{data}
\name{birds}
\alias{birds}
\title{Qualitative data : morphological description of birds}
\format{A data frame with 153 observations on the following 6 variables.

\describe{

\item{\code{V1}}{a numeric vector defining the gender (2 modalities, male or female).}

\item{\code{V2}}{a numeric vector describing the eyebrow stripe (5 modalities).}

\item{\code{V3}}{a numeric vector describing the breast (6 modalities).}

\item{\code{V4}}{a numeric vector describing the stripes (3 modalities).}

\item{\code{V5}}{a numeric vector describing the feathers (5 modalities).}

\item{\code{V6}}{a numeric vector describing the edging (4 modalities).}

}}
\description{
  The dataset contains details on the morphology of birds
  (puffins). Each individual (bird) is described by 6
  qualitative variables. One variable for the gender and 5
  variables giving a morphological description of the
  birds. There is 153 puffins divided in 3 sub-classes:
  dichrous (84 birds), lherminieri (34) and subalaris (35).
}
\examples{
data(birds)
}
\keyword{datasets}

