% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmixmod.R
\docType{data}
\name{car}
\alias{car}
\title{Qualitative data : Car Evaluation}
\format{
A data frame with 1728 observations on the following 6 variables.

\describe{

  \item{\code{buying}}{the buying price (4 modalities: vhigh, high, med, low).}

  \item{\code{maint}}{the price of the maintenance (4 modalities: vhigh, high, med, low).}

  \item{\code{doors}}{the number of doors (4 modalities: 2, 3, 4, 5more).}

  \item{\code{persons}}{the capacity in terms of persons to carry (3 modalities: 2, 4, more).}

  \item{\code{lug_boot}}{the size of luggage boot  (3 modalities: small, med, big).}

  \item{\code{safety}}{the estimated safety of the car (3 modalities: low, med, high).}

  \item{\code{acceptability}}{the car acceptability (4 modalities: unacc, acc, good, vgood).}
}
}
\source{
Creator: Marko Bohanec 
Donors: Marko Bohanec & Blaz Zupan
http://archive.ics.uci.edu/ml/datasets/Car+Evaluation
}
\description{
Car Evaluation Database was derived from a simple hierarchical decision model originally developed for the demonstration of DEX, M. Bohanec, V. Rajkovic: Expert system for decision making.
}
\examples{
  data(car)
}
\keyword{datasets}
