\name{composeModelName}
\alias{composeModelName}
\title{Get the heterogeneous model name using Gaussian and Multnomial model name}
\usage{
  composeModelName(g_modelname, m_modelname)
}
\arguments{
  \item{g_modelname}{Name of Gaussian model}

  \item{m_modelname}{Name of Multinomial model}
}
\value{
  name of heterogeneous model
}
\description{
  Get the heterogeneous model name using Gaussian and
  Multnomial model name
}

