% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixmodResults.R
\name{ellipse}
\alias{ellipse}
\title{Define function to draw an ellipse}
\usage{
ellipse(x, i, j)
}
\arguments{
\item{x}{an object of class [\code{\linkS4class{MixmodResults}}]}

\item{i}{an index of one variable from data}

\item{j}{an index of one variable from data}
}
\description{
Define function to draw an ellipse
}
\keyword{internal}
