% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmixmod.R
\docType{data}
\name{heterodatatest}
\alias{heterodatatest}
\title{Composite data: A testing set}
\format{
A data frame with 100 individuals (rows) and 5 variables (columns).

\describe{

  \item{\code{V1}}{categorical variable with two modalities (1 & 2).}

  \item{\code{V2}}{categorical variable with two modalities (1 & 2).}

  \item{\code{V3}}{categorical variable with two modalities (1 & 2).}

  \item{\code{V4}}{numeric variable.}

  \item{\code{V5}}{numeric variable.}
}
}
\description{
The data set is made up of 5 variables: 3 categorical variables and 2 quantitative variables.
The testing data set has 100 individuals.
}
\examples{
data(heterodatatest)
summary(heterodatatest)
}
\seealso{
\code{\link{heterodatatrain}}
}
\keyword{datasets}
