% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultinomialModel.R, R/GaussianModel.R,
%   R/CompositeModel.R, R/Mixmod.R, R/Strategy.R, R/MixmodCluster.R,
%   R/MixmodLearn.R, R/MixmodPredict.R
\name{initialize,MultinomialModel-method}
\alias{initialize,MultinomialModel-method}
\alias{initialize,GaussianModel-method}
\alias{initialize,CompositeModel-method}
\alias{initialize,Mixmod-method}
\alias{initialize,Strategy-method}
\alias{initialize,MixmodCluster-method}
\alias{initialize,MixmodLearn-method}
\alias{initialize,MixmodPredict-method}
\title{Create an instance of the [\code{\linkS4class{MultinomialModel}}] class using new/initialize.}
\usage{
\S4method{initialize}{MultinomialModel}(
  .Object,
  listModels,
  free.proportions,
  equal.proportions,
  variable.independency,
  component.independency
)

\S4method{initialize}{GaussianModel}(.Object, listModels, family, free.proportions, equal.proportions)

\S4method{initialize}{CompositeModel}(
  .Object,
  listModels,
  free.proportions,
  equal.proportions,
  variable.independency,
  component.independency
)

\S4method{initialize}{Mixmod}(
  .Object,
  data,
  dataType,
  models,
  weight,
  knownLabels,
  xmlIn,
  xmlOut,
  seed,
  trace,
  massiccc
)

\S4method{initialize}{Strategy}(
  .Object,
  algo,
  nbTry,
  initMethod,
  nbTryInInit,
  nbIterationInInit,
  nbIterationInAlgo,
  epsilonInInit,
  epsilonInAlgo,
  seed,
  parameter,
  labels
)

\S4method{initialize}{MixmodCluster}(
  .Object,
  data = NULL,
  nbCluster = NULL,
  dataType = NULL,
  models = NULL,
  strategy = NULL,
  criterion = NULL,
  weight = NULL,
  knownLabels = NULL,
  seed = -1,
  xmlIn = NULL,
  xmlOut = NULL,
  trace = 0,
  massiccc = 0
)

\S4method{initialize}{MixmodLearn}(
  .Object,
  data = NULL,
  knownLabels = NULL,
  dataType = NULL,
  models = NULL,
  criterion = "CV",
  nbCVBlocks = 10,
  weight = NULL,
  seed = -1,
  xmlIn = NULL,
  xmlOut = NULL,
  trace = 0,
  massiccc = 0
)

\S4method{initialize}{MixmodPredict}(
  .Object,
  data,
  classificationRule,
  xmlIn = NULL,
  xmlOut = NULL,
  trace = 0,
  massiccc = 0
)
}
\description{
Initialization method. Used internally in the `Rmixmod' package.

Initialization method. Used internally in the `Rmixmod' package.

Initialization method. Used internally in the `Rmixmod' package.

Initialization method. Used internally in the `Rmixmod' package.

Initialization method. Used internally in the `Rmixmod' package.

Initialization method. Used internally in the `Rmixmod' package.

Initialization method. Used internally in the `Rmixmod' package.

Initialization method. Used internally in the `Rmixmod' package.
}
\seealso{
\code{\link{initialize}}

\code{\link{initialize}}

\code{\link{initialize}}

\code{\link{initialize}}

\code{\link{initialize}}

\code{\link{initialize}}

\code{\link{initialize}}

\code{\link{initialize}}
}
\keyword{internal}
