% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Define function to check an integer}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{a numeric}

\item{tol}{a real which defines the tolerance}
}
\value{
a logical. TRUE if \code{x} is an integer, FALSE otherwise.
}
\description{
Define function to check an integer
}
\keyword{internal}
