% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{isQualitative}
\alias{isQualitative}
\title{Say if a data frame contains only qualitative variables.}
\usage{
isQualitative(x)
}
\arguments{
\item{x}{a vector or a matrix or a data frame}
}
\value{
a boolean
}
\description{
Say if a data frame contains only qualitative variables.
}
