% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{matrix2binary}
\alias{matrix2binary}
\title{Define function to transform a matrix of modalities into matrix of binary data}
\usage{
matrix2binary(x)
}
\arguments{
\item{x}{a matrix}
}
\value{
a matrix of binary data.
}
\description{
Define function to transform a matrix of modalities into matrix of binary data
}
\keyword{internal}
