% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixmodDAResults.R
\docType{class}
\name{MixmodDAResults-class}
\alias{MixmodDAResults-class}
\title{Constructor of [\code{\linkS4class{MixmodDAResults}}] class}
\description{
This is a class to contain results after a discriminant analysis with MIXMOD. Inherits the [\code{\linkS4class{MixmodResults}}] class.
 
\describe{
  \item{CVLabel}{vector of integers containing labels defined by cross validation.}
  \item{CVClassification}{classification table after cross validation.}
  \item{MAPErrorRate}{error rate done by MAP algorithm.}
  \item{MAPClassification}{classification table after MAP algorithm.}
}
}
\examples{
  getSlots("MixmodDAResults")

}
