% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixmodXml.R
\docType{class}
\name{MixmodXmlCheck-class}
\alias{MixmodXmlCheck-class}
\title{Constructor of [\code{\linkS4class{MixmodXmlCheck}}] class}
\description{
This is a class to handle XML files (TODO: describe...)
}
