\name{MixmodXmlInput-class}
\alias{MixmodXmlInput-class}
\title{Constructor of [\code{\linkS4class{MixmodXmlInput}}] class}
\description{
  Internal
}
\details{
  \describe{ \item{family}{Its use is not recommended.} }
}
\examples{
  # new("MixmodXmlInput",file="xml-file", numFormat="humanReadable", conversionOnly=FALSE)

  getSlots("MixmodXmlInput")
}

