\name{mixmodXmlInput}
\alias{mixmodXmlInput}
\title{Internal. It's use is not recommended.}
\alias{mixmodXmlInput}
\usage{
 mixmodXmlInput(\dots)

}
\arguments{
  \item{\dots}{arguments passed to or from methods.}

  \describe{
    Internal. It's use is not recommended.
  }
}
\value{
  a [\code{\linkS4class{MixmodXmlInput}}] object
}
\description{
  This class will contain all the parameters needed by the
  estimation algorithms.
}
\details{
  Internal. It's use is not recommended.

  \describe{
    Internal. It's use is not recommended.
  }
}