% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixmodXml.R
\name{mixmodXmlLoad}
\alias{mixmodXmlLoad}
\title{mixmodXmlLoad}
\usage{
mixmodXmlLoad(xmlFile, numFormat = "humanReadable")
}
\arguments{
\item{xmlFile}{...}

\item{numFormat}{...}
}
\value{
XML output of mixmod methods
}
\description{
TODO: describe...
}
