% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\docType{class}
\name{Model-class}
\alias{Model-class}
\title{Constructor of [\code{\linkS4class{Model}}] class}
\description{
This class defines the Mixmod models.
}
\details{
\describe{
  \item{listModels}{character containing a list of models.}
  \item{free.proportions}{logical to include models with free proportions. Default is TRUE.}
  \item{equal.proportions}{logical to include models with equal proportions. Default is FALSE.}
}
}
\examples{
  getSlots("Model")

}
