\name{nbFactorFromData}
\alias{nbFactorFromData}
\title{Get the number of modalities for each column of a categorical data set}
\usage{
  nbFactorFromData(x)
}
\arguments{
  \item{x}{a vector or a matrix or a data frame containing
  factors}
}
\value{
  a vector containing the number of modalities for each
  column
}
\description{
  Get the number of modalities for each column of a
  categorical data set
}

