% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter.R
\docType{class}
\name{Parameter-class}
\alias{Parameter-class}
\title{Constructor of [\code{\linkS4class{Parameter}}] class}
\description{
This class defines parameters of a Mixture Model.
}
\details{
\describe{
  \item{proportions}{a numeric vector containing proportions of the mixture model.}
}
}
\examples{
  getSlots("Parameter")

}
