% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/MultinomialParameter.R,
%   R/GaussianParameter.R, R/CompositeParameter.R, R/MixmodResults.R,
%   R/Mixmod.R, R/MixmodPredict.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,MultinomialParameter-method}
\alias{summary,GaussianParameter-method}
\alias{summary,CompositeParameter-method}
\alias{summary,MixmodResults-method}
\alias{summary,Mixmod-method}
\alias{summary,MixmodPredict-method}
\title{Produce result summaries of a Rmixmod class}
\usage{
\S4method{summary}{MultinomialParameter}(object, ...)

\S4method{summary}{GaussianParameter}(object, ...)

\S4method{summary}{CompositeParameter}(object, ...)

\S4method{summary}{MixmodResults}(object, ...)

\S4method{summary}{Mixmod}(object, ...)

\S4method{summary}{MixmodPredict}(object, ...)
}
\arguments{
\item{object}{An object (???)}

\item{...}{further arguments passed to or from other methods}
}
\value{
NULL. Summaries to standard out.
}
\description{
Produce result summaries of a Rmixmod class
}
\examples{
  data(geyser)
  xem <- mixmodCluster(geyser,3)
  summary(xem)
  summary(xem["bestResult"])
  summary(xem["bestResult"]["parameters"])

}
\seealso{
\code{\link{summary}}
}
