\docType{data}
\name{birds}
\alias{birds}
\title{Qualitative data : morphological description of birds}
\format{A data frame with 69 observations on the following 6 variables.

\describe{

  \item{\code{gender}}{a numeric vector defining the gender (2 modalities, male or female).}

  \item{\code{eyebrow}}{a numeric vector describing the eyebrow stripe (5 modalities).}

  \item{\code{breast}}{a numeric vector describing the breast (6 modalities).}

  \item{\code{stripes}}{a numeric vector describing the stripes (3 modalities).}

  \item{\code{feathers}}{a numeric vector describing the feathers (5 modalities).}

  \item{\code{edging}}{a numeric vector describing the edging (4 modalities).}

}}
\description{
  The dataset contains details on the morphology of birds
  (puffins). Each individual (bird) is described by 6
  qualitative variables. One variable for the gender and 5
  variables giving a morphological description of the
  birds. There is 69 puffins divided in 2 sub-classes:
  lherminieri (34) and subalaris (35).
}
\examples{
data(birds)
}
\keyword{datasets}

