\docType{methods}
\name{[}
\alias{[}
\alias{[,GaussianParameter,ANY,numeric-method}
\alias{[,MixmodClustering,ANY,numeric-method}
\alias{[,MultinomialParameter,ANY,numeric-method}
\alias{[<-,GaussianParameter-method}
\alias{[<-,GaussianParameter,ANY-method}
\alias{[<-,MixmodClustering-method}
\alias{[<-,MixmodClustering,ANY-method}
\alias{[<-,MultinomialParameter-method}
\alias{[<-,MultinomialParameter,ANY-method}
\title{Extract parts of a [\code{\linkS4class{GaussianParameter}}] class}
\arguments{
  \item{x}{object from which to extract element(s) or in
  which to replace element(s).}

  \item{i}{the name of the element we want to extract or
  replace.}

  \item{j}{if the element designing by i is complex, j
  specifying elements to extract or replace.}

  \item{drop}{For matrices and arrays.  If TRUE the result
  is coerced to the lowest possible dimension (see the
  examples).  This only works for extracting elements, not
  for the replacement.  See drop for further details.}
}
\description{
  Extract parts of a
  [\code{\linkS4class{GaussianParameter}}] class

  Replace names of [\code{\linkS4class{GaussianParameter}}]
  class

  Extract parts of a
  [\code{\linkS4class{MixmodClustering}}] class

  Replace names of [\code{\linkS4class{MixmodClustering}}]
  class

  Extract parts of a
  [\code{\linkS4class{MultinomialParameter}}] class

  Replace names of
  [\code{\linkS4class{MultinomialParameter}}] class
}

