% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/MultinomialModel.R, R/GaussianModel.R, R/CompositeModel.R, R/MultinomialParameter.R, R/GaussianParameter.R, R/CompositeParameter.R, R/MixmodResults.R, R/Strategy.R, R/MixmodCluster.R, R/MixmodLearn.R, R/MixmodPredict.R
\docType{methods}
\name{[}
\alias{[}
\alias{[,CompositeModel-method}
\alias{[,CompositeParameter-method}
\alias{[,GaussianModel-method}
\alias{[,GaussianParameter-method}
\alias{[,MixmodCluster-method}
\alias{[,MixmodLearn-method}
\alias{[,MixmodPredict-method}
\alias{[,MixmodResults-method}
\alias{[,MultinomialModel-method}
\alias{[,MultinomialParameter-method}
\alias{[,Strategy-method}
\alias{[<-,CompositeModel-method}
\alias{[<-,CompositeParameter-method}
\alias{[<-,GaussianModel-method}
\alias{[<-,GaussianParameter-method}
\alias{[<-,MixmodCluster-method}
\alias{[<-,MixmodLearn-method}
\alias{[<-,MultinomialModel-method}
\alias{[<-,MultinomialParameter-method}
\alias{[<-,Strategy-method}
\title{Extract parts of a Rmixmod class}
\usage{
\S4method{[}{MultinomialModel}(x, i, j, drop)

\S4method{[}{MultinomialModel}(x, i, j) <- value

\S4method{[}{GaussianModel}(x, i, j, drop)

\S4method{[}{GaussianModel}(x, i, j) <- value

\S4method{[}{CompositeModel}(x, i, j, drop)

\S4method{[}{CompositeModel}(x, i, j) <- value

\S4method{[}{MultinomialParameter}(x, i, j, drop)

\S4method{[}{MultinomialParameter}(x, i, j) <- value

\S4method{[}{GaussianParameter}(x, i, j, drop)

\S4method{[}{GaussianParameter}(x, i, j) <- value

\S4method{[}{CompositeParameter}(x, i, j, drop)

\S4method{[}{CompositeParameter}(x, i, j) <- value

\S4method{[}{MixmodResults}(x, i, j, drop)

\S4method{[}{Strategy}(x, i, j, drop)

\S4method{[}{Strategy}(x, i, j) <- value

\S4method{[}{MixmodCluster}(x, i, j, drop)

\S4method{[}{MixmodCluster}(x, i, j) <- value

\S4method{[}{MixmodLearn}(x, i, j, drop)

\S4method{[}{MixmodLearn}(x, i, j) <- value

\S4method{[}{MixmodPredict}(x, i, j, drop)
}
\arguments{
\item{x}{object from which to extract element(s) or in which to replace element(s).}

\item{i}{the name of the element we want to extract or replace.}

\item{j}{if the element designing by i is complex, j specifying elements to extract or replace.}

\item{drop}{For matrices and arrays.  If TRUE the result is coerced to the lowest possible dimension (see the examples).  This only works for extracting elements, not for the replacement.  See drop for further details.}

\item{value}{TODO...}
}
\description{
Extract parts of a Rmixmod class
}

