\name{MultinomialModel-class}
\alias{MultinomialModel-class}
\title{Constructor of [\code{\linkS4class{MultinomialModel}}] class}
\description{
  This class defines a multinomial Model. Inherits the
  [\code{\linkS4class{Model}}] class.
}
\details{
  \describe{

  \item{variable.independency}{logical}

  \item{component.independency}{logical}

  }
}
\examples{
new("MultinomialModel")
}
\author{
  Remi Lebret \email{remi.lebret@math.univ-lille1.fr}
}

