% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{vector2binary}
\alias{vector2binary}
\title{Define function to transform a vector of modalities into matrix of binary data}
\usage{
vector2binary(x)
}
\arguments{
\item{x}{a numeric}
}
\value{
a matrix of binary data.
}
\description{
Define function to transform a vector of modalities into matrix of binary data
}
\keyword{internal}
